requirements.txt
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-bigquery==2.6.1
google-cloud-core==1.5.0
google-crc32c==1.1.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
httplib2==0.18.1
idna==2.10
numpy==1.19.5
pandas==1.2.0
proto-plus==1.13.0
protobuf==3.14.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
rsa==4.7
six==1.15.0
tqdm==4.56.0
Unidecode==1.1.2
uritemplate==3.0.1
urllib3==1.26.2
xlrd==2.0.1
